we sure love merge commits
This reverts commit 3041d7d8e03c007023a56197204d691595abe55c, reversing
changes made to b4de52d39b9d3f73c55aee79d575df94b3647092.

+++ src/main/java/frc/robot/subsystems/Elevator.java
@@ -37,7 +37,7 @@ public class Elevator extends SubsystemBase {
 
     public void setSpeed(double speed) {
         // Encoder Clicks
-        double encoder = getEncoderPosition();
+        int encoder = getEncoderPosition();
         boolean limitReached = false;
 
         // If Going down set a lower limit of a encoder value of the constant lower limit
@@ -56,7 +56,7 @@ public class Elevator extends SubsystemBase {
         elevator.set(ControlMode.PercentOutput, speed);
     }
 
-    public double getEncoderPosition() {
+    public int getEncoderPosition() {
         return elevator.getSelectedSensorPosition();
     }
